ggDoubleHeat: Data visualization for two sources

ggDoubleHeat, which is a ggplot2 extension, provides visualization for a reformed heatmap. Instead of facetting heatmaps by data sources, they can be combined together for making one single heatmap, generated by geom_heat_*() functions built in the package. Prior to using the package, users should load ggplot2.

ggDoubleHeat functions

All functions in the package are named as geom_heat_*(), making the naming convention consistent. Also, the arguments for functions are relatively similar, although with slight variations due to where a specific argument will connect to the position of the rendering plot. Users should reference the documentation and possibly run examples presented in the help file when trying to understand what each argument means visually.

See also

Author

Maintainer: Youzhi Yu yuyouzhi666@icloud.com

Authors:

Trent Buskirk [thesis advisor]